requirements.txt
streamlit
pandas
gspread
oauth2client
fpdf2
python-dotenv